refactor: Rename navbar to temp_navbar
