+++ src/pages/petCardPage/PetCardPage.tsx
@@ -5,7 +5,7 @@ import { MainContainer } from "@shared/ui/mainContainer";
 import { SearchPet } from "@features/pet/searchPet";
 import { Button } from "@shared/ui/button";
 import { ADD_PET_CARD } from "@app/router/consts";
-import { Navbar } from "@widgets/navbar";
+import { Navbar } from "@/widgets/temp_navbar";
 import { FilterPets } from "@/features/pet/filter";
 import { PetCardDetail } from "@widgets/petCardDetail/ui/PetCardDetail";
