+++ src/widgets/temp_navbar/ui/navbar.module.scss
